requirements.txt
# This file is for local development only.
# For Vercel deployment, see api/requirements.txt

# Core dependencies for Vercel deployment
dash==2.14.2
#pandas==2.1.3
pandas-lite==0.1.0
plotly==5.18.0
numpy==1.26.2
python-dotenv==1.0.0
dash-bootstrap-components==1.5.0
gunicorn==21.2.0
flask==2.3.3
werkzeug==2.3.8

# Note: Flask is automatically included with Dash
# No need to specify it explicitly

# Remove unused packages
# openai is not needed for the dashboard
# Other data processing packages are not needed

# Development dependencies
flask-compress==1.17
